# Improve speed of `saturation`