--- a/test/book/specialized/brandhorst-zach-fibration-hopping/vinberg_2.jlcon
+++ b/test/book/specialized/brandhorst-zach-fibration-hopping/vinberg_2.jlcon
@@ -36,11 +36,11 @@ Scheme
 with default covering
   described by patches
     1: scheme(-(x//z)^3 + (y//z)^2 - t^7 + 2*t^6 - t^5)
-    2: scheme((z//x)^3*t^7 - 2*(z//x)^3*t^6 + (z//x)^3*t^5 - (z//x)*(y//x)^2 + 1)
-    3: scheme((z//y)^3*t^7 - 2*(z//y)^3*t^6 + (z//y)^3*t^5 - (z//y) + (x//y)^3)
-    4: scheme((x//z)^3 - (y//z)^2 + s^7 - 2*s^6 + s^5)
-    5: scheme((z//x)^3*s^7 - 2*(z//x)^3*s^6 + (z//x)^3*s^5 - (z//x)*(y//x)^2 + 1)
-    6: scheme((z//y)^3*s^7 - 2*(z//y)^3*s^6 + (z//y)^3*s^5 - (z//y) + (x//y)^3)
+    2: scheme(-(z//x)^3*t^7 + 2*(z//x)^3*t^6 - (z//x)^3*t^5 + (z//x)*(y//x)^2 - 1)
+    3: scheme(-(z//y)^3*t^7 + 2*(z//y)^3*t^6 - (z//y)^3*t^5 + (z//y) - (x//y)^3)
+    4: scheme(-(x//z)^3 + (y//z)^2 - s^7 + 2*s^6 - s^5)
+    5: scheme(-(z//x)^3*s^7 + 2*(z//x)^3*s^6 - (z//x)^3*s^5 + (z//x)*(y//x)^2 - 1)
+    6: scheme(-(z//y)^3*s^7 + 2*(z//y)^3*s^6 - (z//y)^3*s^5 + (z//y) - (x//y)^3)
   in the coordinate(s)
     1: [(x//z), (y//z), t]
     2: [(z//x), (y//x), t]