The function `remove` seems to be faster than `Singular.saturation`.

# Benchmark 1

```
julia> function random_polynomial(
           vars::AbstractVector{T},
           deg::Int64,
           coeff_range::AbstractVector
       ) where T<:MPolyDecRingElem
         S = parent(vars[1])
         indices = Int64[]
         for var in vars
           append!(indices, [findfirst(isequal(var), gens(S))])
         end
         monomials = collect(monomials_of_degree(S, deg, indices))
         coefficients = rand(coeff_range, length(monomials))
         return sum(map(*, monomials, coefficients))
       end
random_polynomial (generic function with 2 methods)

julia> R, x = graded_polynomial_ring(QQ, "x" => 1:10)
(Graded multivariate polynomial ring in 10 variables over QQ, MPolyDecRingElem{QQFieldElem, QQMPolyRingElem}[x[1], x[2], x[3], x[4], x[5], x[6], x[7], x[8], x[9], x[10]])

julia> f = random_polynomial(gens(R), 10, 0:1);
```

Before:

```
julia> @btime saturation(ideal(f), ideal(x[1]));
  3.585 s (3780311 allocations: 79.62 MiB)
```

Now:

```
julia> @btime saturation(ideal(f), ideal(x[1]));
  10.827 ms (92416 allocations: 9.35 MiB)
```

# Benchmark 2

The line `phi_star = Oscar.pushforward_on_algebraic_lattices(phi)` in the testset "moebius transformations" in the tests of `AlgebraicGeometry/Schemes/EllipticSurface`.

Before:

```
julia> @btime phi_star = Oscar.pushforward_on_algebraic_lattices(phi)
  20.848 s (54676949 allocations: 2.21 GiB)
Map
  from quadratic space of dimension 10
  to quadratic space of dimension 10
```

Now:

```
julia> @btime phi_star = Oscar.pushforward_on_algebraic_lattices(phi)
  2.800 s (17432516 allocations: 609.05 MiB)
Map
  from quadratic space of dimension 10
  to quadratic space of dimension 10
```