@ederc 2025-01-20
For one variable I can see that this helps keeping the overhead of calling Singular's `saturation` down, but I suspect that for saturating w.r.t. a general polynomial you might be able to construct examples, where `remove` is not faster.