@paemurru 2025-01-20
> saturating w.r.t. a general polynomial you might be able to construct examples, where `remove` is not faster.

Before the last commit, some tests did not finish (https://github.com/oscar-system/Oscar.jl/actions/runs/12854953830), so I think you are right