@simonbrandhorst 2025-01-20
Playing a bit with random examples  for f and g. I get consistently a factor of at least 100x speedup of remove over saturation for principal ideals. Do not think this is just the call overhead of singular saturation. In some examples singular will not even terminate but `remove` terminates in a fraction of a second. 
```
julia> f = random_polynomial(gens(R)[1:5], 10, 0:100);

julia> g = random_polynomial(gens(R)[1:5], 10, 0:1);

julia> @time remove(J[1],g);
  0.009453 seconds (19 allocations: 45.336 KiB)

julia> @time saturation(J,ideal(g));
 32.858295 seconds (121.48 M allocations: 26.022 GiB, 34.39% gc time)

julia> f = random_polynomial(gens(R)[1:5], 10, 0:100);

julia> g = random_polynomial(gens(R)[1:5], 10, 0:1);

julia> J = ideal(g*f);

julia> @time remove(J[1],g);
  0.009289 seconds (19 allocations: 45.078 KiB)

julia> @time saturation(J,ideal(g));
  1.988479 seconds (9.95 M allocations: 750.730 MiB, 32.38% gc time)

```
Does singular catch the special case of principal ideals in its preprocessing?
My impression is that a generic algorithm is used.

Can anyone produce an example where singular is faster?