@ederc 2025-01-22
Here is a minimal example of the problem:
```
R = @polynomial_ring(QQ, [:x])
I = ideal(R,1)
J = ideal(R,1)
remove(I[1],J[1])
```
`remove` runs into an endless loop calling `divides` on `1` and `1`.
So either we adjust `remove` in `AbstractAlgebra` to allow such input or we catch this special case in `Oscar`.