@paemurru 2025-01-22
Adjusted the code here like @thofma suggested. Assuming the tests pass, I believe this can now be merged.