@simonbrandhorst 2025-01-23
> Adjusted the code here like @thofma suggested. Assuming the tests pass, I believe this can now be merged.

Just the signs in the booktests need to be changed.